+++ GuaranteedRate.Sextant.Config/IJsonEncompassConfig.cs
