+++ GuaranteedRate.Sextant.Config/JsonEncompassConfig.cs
