+++ GuaranteedRate.Sextant.Logging.Elasticsearch/packages.config
@@ -1,0 +1,7 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Elasticsearch.Net" version="6.0.0-alpha1" targetFramework="net452" />
+  <package id="EncompassSDK.Complete" version="17.2.0.2" targetFramework="net452" />
+  <package id="NEST" version="5.4.0" targetFramework="net452" />
+  <package id="Newtonsoft.Json" version="10.0.1" targetFramework="net452" />
+</packages>
