+++ GuaranteedRate.Sextant.Logging.Loggly/Class1.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace GuaranteedRate.Sextant.Logging.Loggly
+{
+    public class Class1
+    {
+    }
+}
