+++ GuaranteedRate.Sextant.Logging/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="EncompassSDK.Complete" version="17.2.0.2" targetFramework="net452" />
+</packages>
