+++ GuaranteedRate.Sextant/EncompassUtils/FieldUtils.cs
@@ -5,7 +5,6 @@ using EllieMae.Encompass.Collections;
 using EllieMae.Encompass.Reporting;
 using System;
 using System.Collections.Generic;
-using GuaranteedRate.Sextant.Loggers;
 
 namespace GuaranteedRate.Sextant.EncompassUtils
 {
@@ -765,7 +764,7 @@ namespace GuaranteedRate.Sextant.EncompassUtils
                 }
                 else
                 {
-                    Loggly.Error("FieldUtils", $"Duplicate field Ids ({fieldDescriptor.FieldID}) found in GetAllFieldDescriptors. "+ 
+                    Logger.Error("FieldUtils", $"Duplicate field Ids ({fieldDescriptor.FieldID}) found in GetAllFieldDescriptors. "+ 
                         $"Existing value: {fieldsAndDescriptions[fieldDescriptor.FieldID]}, New value: {fieldDescriptor.Description}");
                 }
             }
