+++ GuaranteedRate.SextantTests/packages.config
@@ -1,6 +1,8 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="Elasticsearch.Net" version="6.0.0-alpha1" targetFramework="net45" />
   <package id="EncompassSDK.Complete" version="17.2.0.2" targetFramework="net45" />
+  <package id="NEST" version="5.4.0" targetFramework="net45" />
   <package id="Newtonsoft.Json" version="10.0.2" targetFramework="net45" />
   <package id="NUnit" version="2.6.4" targetFramework="net45" />
 </packages>
